Updated With Minimum Follower Requirements for Followers
Twitter can occasionally present a bug in which a user's followers are listed as "1" rather than an actual value. We see this with EthiopiaLiveUpd, where one tweet in the dataset for this user only had one follower even though before and after that tweet, that user had ~12k followers. To address this issue, I put in a restriction that a user must have >= 20 followers to be checked in the follower-gain metrics.